Refactoring paging components to extract common default sorting options
