+++ src/app/core/admin/end-user-agreement/admin-list-euas.component.ts
@@ -144,13 +144,7 @@ export class AdminListEuasComponent extends PagingComponent implements OnDestroy
 		if (cachedFilter.paging) {
 			this.pagingOpts = cachedFilter.paging;
 		} else {
-			this.pagingOpts = new PagingOptions();
-
-			const defaultSort = this.headers.find((header: any) => header.default);
-			if (null != defaultSort) {
-				this.pagingOpts.sortField = defaultSort.sortField;
-				this.pagingOpts.sortDir = defaultSort.sortDir;
-			}
+			this.setDefaultPagingOptions();
 		}
 	}
